fix re-render component in component bug
This error appears in console as
"Cannot update a component (`ConnectFunction`) while rendering a different component (`AdminLayout`)"
It is because while the Layout component is rendering, we dispatch changes to the component again. Hence, we need to use redux instead to fetch different menus directly during load instead of dispatching and replacing the menu.